Renaming the Valkyrie::Indexer shared spec to specifically use Valkyrie::Persistence::Solr::Indexer; Isolating the Indexer for the shared test suite
