+++ lib/valkyrie/specs/shared_specs.rb
@@ -7,4 +7,4 @@ require 'valkyrie/specs/shared_specs/storage_adapter.rb'
 require 'valkyrie/specs/shared_specs/change_set_persister.rb'
 require 'valkyrie/specs/shared_specs/file.rb'
 require 'valkyrie/specs/shared_specs/change_set.rb'
-require 'valkyrie/specs/shared_specs/indexer.rb'
+require 'valkyrie/specs/shared_specs/solr_indexer.rb'
